Moving tests out of app files and into their own files
