+++ slam_sim/src/globals.cpp
@@ -1,0 +1,16 @@
+////////////////////////////////////////////////////
+// Copyright (c) Gaia Platform LLC
+//
+// Use of this source code is governed by the MIT
+// license that can be found in the LICENSE.txt file
+// or at https://opensource.org/licenses/MIT.
+////////////////////////////////////////////////////
+
+#include <cstdint>
+
+namespace slam_sim
+{
+
+int32_t g_quit = 0;
+
+}
